backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-socketio[asgi]
pydantic
SQLAlchemy
pydantic-settings
pydantic[email]
anyio
starlette